xone: dongle: Warn if counterfeit MAC address detected
It seems there are a number of counterfeit xbox wireless
adapters out there, some of which seem to be unsurprisingly
provisioned with the same MAC address. This will create
issues if two of them are used in the same RF environment.
Detect such suspicous MAC addresses and throw a warning,
suggesting the user pass in override_mac= for deconflicton.

Signed-off-by: evilwombat <2313235+evilwombat@users.noreply.github.com>